Disable two way nest, first close mass leak.

+++ cases/drycbl_nesting/drycbl_settings.jl
@@ -7,7 +7,7 @@ end
 
 
 ## Default float type.
-float_type = Float64
+float_type = Float32
 
 
 ## Settings domain 1.
